Split the source code file in 4 files by functionality to improve code modularity
